Moved App file into its own component folder
